requirements.txt
# Core dependencies
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
orjson==3.11.4
packaging==25.0
pydantic==2.12.4
pydantic_core==2.41.5
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
zstandard==0.25.0

# LangChain dependencies
langchain==1.0.2
langchain-core==1.0.4
langchain-openai==1.0.2
langchain-postgres==0.0.15
langchain-text-splitters==1.0.0
langsmith==0.4.42

# OpenAI
openai==2.7.2
tiktoken==0.12.0

# PostgreSQL and Vector dependencies
asyncpg==0.30.0
pgvector==0.3.6
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
psycopg2-binary==2.9.10
SQLAlchemy==2.0.43
greenlet==3.2.4

# PDF processing
pypdf==6.1.2
filetype==1.2.0

# Environment management
python-dotenv==1.1.1

# Numpy for vector operations
numpy==2.3.2
